@@ -1,6 +1,6 @@
 {
   "name": "web-switcher",
-  "version": "0.4.0",
+  "version": "0.5.0",
   "private": true,
   "description": "Build your UserScript with webpack",
   "license": "MIT",
@@ -13,9 +13,9 @@
     "email": "krysl@qq.com"
   },
   "scripts": {
-    "anylize": "npm_config_report=true npm run build",
-    "build": "webpack --mode production --config config/webpack.config.production.js",
-    "dev": "webpack --mode development --config config/webpack.config.dev.js",
+    "anylize": "cross-env npm_config_report=true npm run build",
+    "build": "cross-env TS_NODE_PROJECT=\"config/tsconfig-for-webpack-config.json\"  webpack --mode production --config config/webpack.config.production.ts",
+    "dev": "cross-env TS_NODE_PROJECT=\"config/tsconfig-for-webpack-config.json\"  webpack --mode development --config config/webpack.config.dev.ts",
     "lint": "eslint src -f visualstudio",
     "preversion": "npm run lint",
     "postversion": "git push --follow-tags"
@@ -34,26 +34,28 @@
     "element-plus": "^1.0.2-beta.32",
     "jquery": "3.5.1",
     "loglevel": "^1.7.1",
-    "vue": "^3.0.5",
+    "vue": "^3.0.6",
     "vue-class-component": "^8.0.0-rc.1",
     "vuex": "^4.0.0"
   },
   "devDependencies": {
-    "@babel/core": "^7.12.17",
-    "@babel/plugin-proposal-class-properties": "^7.12.13",
-    "@babel/plugin-proposal-decorators": "^7.12.13",
-    "@babel/preset-env": "^7.12.17",
+    "@babel/core": "^7.13.1",
+    "@babel/plugin-proposal-class-properties": "^7.13.0",
+    "@babel/plugin-proposal-decorators": "^7.13.5",
+    "@babel/preset-env": "^7.13.5",
     "@types/estree": "0.0.46",
     "@types/jquery": "^3.5.5",
+    "@types/node": "^14.14.31",
     "@types/tampermonkey": "^4.0.0",
-    "@typescript-eslint/eslint-plugin": "4.15.1",
-    "@typescript-eslint/parser": "4.15.1",
+    "@types/webpack": "^4.41.26",
+    "@typescript-eslint/eslint-plugin": "4.15.2",
+    "@typescript-eslint/parser": "4.15.2",
     "@vue/babel-helper-vue-jsx-merge-props": "^1.2.1",
     "@vue/babel-plugin-jsx": "^1.0.3",
-    "@vue/compiler-sfc": "^3.0.5",
+    "@vue/compiler-sfc": "^3.0.6",
     "@vue/composition-api": "1.0.0-rc.2",
-    "@vue/runtime-core": "^3.0.5",
-    "@vue/runtime-dom": "^3.0.5",
+    "@vue/runtime-core": "^3.0.6",
+    "@vue/runtime-dom": "^3.0.6",
     "babel-helper-vue-jsx-merge-props": "^2.0.3",
     "babel-loader": "^8.2.2",
     "babel-plugin-import": "^1.13.3",
@@ -74,12 +76,13 @@
     "style-loader": "2.0.0",
     "terser-webpack-plugin": "^5.1.1",
     "ts-loader": "8.0.17",
-    "typescript": "4.1.5",
+    "ts-node": "^9.1.1",
+    "typescript": "^4.2.2",
     "url-loader": "^4.1.1",
     "userscript-metadata-webpack-plugin": "0.0.6",
     "vue-loader": "^16.1.2",
     "vue-style-loader": "^4.1.2",
-    "webpack": "5.23.0",
+    "webpack": "5.24.2",
     "webpack-bundle-analyzer": "^4.4.0",
     "webpack-cli": "4.5.0",
     "webpack-livereload-plugin": "2.3.0",
