@@ -1,25 +1,29 @@
-import * as elternalsCSS from '../../config/element-plus-elternals-css';
-import * as pkg from '../../package.json';
+import { elternalsCSS } from '../style/element-plus-elternals-css';
+import pkg from '../../package.json';
 
 async function loadElementPlusStyle() {
   const globalVars = new Map<string, string>();
+  const version = pkg.dependencies['element-plus'];
   const ret = elternalsCSS.map(async (name) => {
-    const globalName = `theme_chalk_${name.replace(/-/g, '_')}`;
-    const style = await GM_getResourceText(globalName).replaceAll(
+    const globalName = `theme_chalk_${(typeof name === 'string'
+      ? name
+      : name.name
+    ).replace(/-/g, '_')}`;
+    let style = await GM_getResourceText(globalName).replaceAll(
       'url(',
-      `url(https://cdn.jsdelivr.net/npm/element-plus@${pkg.dependencies['element-plus']}/lib/theme-chalk/`
+      `url(https://cdn.jsdelivr.net/npm/element-plus@${version}/lib/theme-chalk/`
     );
+    if (typeof name !== 'string' && name.patch !== undefined) {
+      for (const key in name.patch) {
+        style = style.replaceAll(key, name.patch[key]);
+      }
+    }
     window[globalName] = style;
+    GM_addStyle(style);
     globalVars.set(globalName, window[globalName]);
+    // console.debug(`GM_addStyle: ${globalVars.get(globalName)}`);
   });
   await Promise.all(ret);
-
-  GM_addStyle(
-    await GM_getResourceText('theme_chalk_index').replaceAll(
-      'url(',
-      `url(https://cdn.jsdelivr.net/npm/element-plus@${pkg.dependencies['element-plus']}/lib/theme-chalk/`
-    )
-  );
   return globalVars;
 }
 
